# contrato/openapi.yaml
openapi: 3.0.3
info:
  title: Motor do fechamento — contrato comum do spike
  version: "1.0.0"
  description: |
    Contrato que as DUAS spikes (.NET 10 e Python) implementam exatamente igual.
    Semântica: docs/semantica-do-motor.md. O verificador chama só estas rotas.
    Nenhuma rota extra é exigida; rotas extras não são usadas pelo verificador.
    Sem autenticação no spike.
servers:
  - url: http://localhost:8081
    description: spike .NET
  - url: http://localhost:8082
    description: spike Python
paths:
...
  /competencias/abrir:
    post:
      summary: AbrirCompetencias (semântica §5)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AbrirCompetenciasRequest" }
      responses:
        "200":
          description: Casos criados ou já existentes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AbrirCompetenciasResponse" }
        "422": { $ref: "#/components/responses/Invalido" }
  /fatos:
    post:
      summary: PublicarFato (semântica §6)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FatoRequest" }
      responses:
        "200":
          description: Fato gravado ou sem mudança
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FatoResponse" }
        "422": { $ref: "#/components/responses/Invalido" }
...
components:
...
  responses:
...
    Invalido:
      description: Corpo inválido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Erro" }
  schemas:
    Erro:
      type: object
      required: [erro, mensagem]
      properties:
        erro: { type: string, description: "código curto, ex.: 'entregavel_nao_pronto'" }
        mensagem: { type: string }
    Empresa:
      type: object
      additionalProperties: false
      required: [titular_id, regime, tem_folha, tem_prolabore, tem_taxa_municipal, filial, carteira, municipio]
      properties:
        titular_id: { type: string, minLength: 1, maxLength: 64 }
        regime: { type: string, enum: [SN, LP] }
        tem_folha: { type: boolean }
        tem_prolabore: { type: boolean }
        tem_taxa_municipal: { type: boolean }
        filial: { type: boolean }
        carteira: { type: string, minLength: 1 }
        municipio: { type: string, minLength: 1 }
    AbrirCompetenciasRequest:
      type: object
      additionalProperties: false
      required: [competencia, ator, empresas]
      properties:
        competencia: { type: string, pattern: "^[0-9]{4}(0[1-9]|1[0-2])$" }
        ator: { type: string, minLength: 1 }
        empresas:
          type: array
          minItems: 1
          maxItems: 10000
          items: { $ref: "#/components/schemas/Empresa" }
    AbrirCompetenciasResponse:
      type: object
      required: [criados, existentes]
      properties:
        criados:
          type: array
          items:
            type: object
            required: [titular_id, caso_id]
            properties:
              titular_id: { type: string }
              caso_id: { type: string, format: uuid }
        existentes:
          type: array
          items:
            type: object
            required: [titular_id, caso_id]
            properties:
              titular_id: { type: string }
              caso_id: { type: string, format: uuid }
    FatoRequest:
      type: object
      additionalProperties: false
      required: [titular_id, competencia, tipo, fonte, observado_em]
      properties:
        titular_id: { type: string, minLength: 1, maxLength: 64 }
        competencia: { type: string, pattern: "^[0-9]{4}(0[1-9]|1[0-2])$" }
        tipo: { type: string, minLength: 1 }
        tributo: { type: string, default: "" }
        valor_centavos: { type: integer, format: int64, nullable: true }
        payload: { type: object, default: {} }
        fonte: { type: string, minLength: 1 }
        origem_ref: { type: string, nullable: true }
        observado_em: { type: string, format: date-time }
    FatoResponse:
      type: object
      required: [fato_id, versao, efeito]
      properties:
        fato_id: { type: string, format: uuid, description: "id da versão vigente após a operação" }
        versao: { type: integer }
        efeito: { type: string, enum: [novo, nova_versao, sem_mudanca] }
...